Change some comments, replace var x = by x:= inside functions
